Fixed .gitignore issue where deeply nested XML files are excluded
Due to a limitation with Git you can't ignore a folder and then
negate files/folders within the ignored folder beyond a certain
depth. Instead, all the contents within the ignored folder are
ignored so the full depth is traversed. Then specific files/folders
within the ignored folder are negated.