Updated all templates to remove asset loading error messages
These error messages came from an SDK update that provides an error
message when an asset does not load. All the templates used a
generic, default base file name on each asset, but the file doesn't
actually exist. They relied on the silent failure during
initialization because a valid file name is later set in the code.
To fix this, the default base file name is just set to an empty
string.